+++ tests/test_models.py
@@ -19,4 +19,4 @@ def test___check_user___password___check_password_equals_false(new_user):
 
 def test___init_user___email_as_number___raise_error():
     with pytest.raises(TypeError):
-        activity_logger.models.User()  # pylint: disable=no-value-for-parameter
+        activity_logger.models.org.User()  # pylint: disable=no-value-for-parameter
